feat: register, login and profile endpoints

+++ libs/api/feature-user/src/lib/api-feature-user.controller.ts
@@ -2,13 +2,18 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   HttpCode,
   Param,
   Post,
+  Request,
+  UseGuards,
 } from '@nestjs/common';
 import { ApiFeatureUserService } from './api-feature-user.service';
 import { CreateUserDto } from './dtos/user.dto';
+import { SignInDto } from './dtos/sign-in.dto';
+import { AuthGuard } from './api-feature-user.guard';
 
 @Controller({ path: 'users' })
 export class ApiFeatureUserController {
@@ -19,19 +24,38 @@ export class ApiFeatureUserController {
     return this.apiFeatureUserService.getAll();
   }
 
+  @UseGuards(AuthGuard)
+  @Get('profile')
+  getProfile(@Request() req: any) {
+    return req.user;
+  }
+
   @Get(':id')
   getOne(@Param('id') id: string) {
     return this.apiFeatureUserService.getById(id);
   }
 
-  @Post('')
-  create(@Body() data: CreateUserDto) {
+  @Post('register')
+  register(@Body() data: CreateUserDto) {
     return this.apiFeatureUserService.create(data);
   }
 
+  @Post('login')
+  @HttpCode(200)
+  signIn(@Body() signInDto: SignInDto) {
+    return this.apiFeatureUserService.signIn(
+      signInDto.email,
+      signInDto.password,
+    );
+  }
+
+  @UseGuards(AuthGuard)
   @Delete(':id')
   @HttpCode(204)
-  delete(@Param('id') id: string) {
+  delete(@Request() req: any, @Param('id') id: string) {
+    if (req.user.sub !== id) {
+      throw new ForbiddenException();
+    }
     return this.apiFeatureUserService.delete(id);
   }
 }
